endpoints/assistant/urls.py
from django.urls import path
from rest_framework_simplejwt.views import TokenRefreshView
from .views import RegisterView, LoginView, PasswordResetView, PasswordResetConfirmView, MedicalAssistantAPI
from django.conf import settings
from django.conf.urls.static import static


urlpatterns = [
    # Authentication Endpoints
    path('', RegisterView.as_view(), name='register'),
    path('login/', LoginView.as_view(), name='login'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),

    # Password Reset Endpoints
    path('password-reset/', PasswordResetView.as_view(), name='password_reset'),
    path('password-reset/confirm/', PasswordResetConfirmView.as_view(), name='password_reset_confirm'),
    path('medical-assistant/', MedicalAssistantAPI.as_view(), name='assistant'),
    

    # urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
]
urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)